requirements.txt
selenium==4.6.1
requests==2.28.1
python-dotenv==0.21.0
pillow==9.3.0